requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
albumentations=1.3.1=pypi_0
appdirs=1.4.4=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=py310hdda0065_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.0=pypi_0
docker-pycreds=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
fonttools=4.43.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.9.2=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.37=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.17.3=pypi_0
idna=3.4=py310h06a4308_0
imageio=2.31.4=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py310h06a4308_0
joblib=1.2.0=py310h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lightning-utilities=0.9.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py310h7f8727e_0
matlabengine=9.14.3=pypi_0
matplotlib=3.8.0=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
numpy=1.26.0=py310h5f9d8c6_0
numpy-base=1.26.0=py310hb5e798b_0
opencv-python=4.8.1.78=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.11=h7f8727e_2
packaging=23.2=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.4.0=py310h6a678d5_1
pip=23.2.1=py310h06a4308_0
protobuf=4.24.3=pypi_0
psutil=5.9.0=py310h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.3.1=py310_torch_2.0.0_cu118
pymupdf=1.23.5=pypi_0
pymupdfb=1.23.5=pypi_0
pyopenssl=23.2.0=py310h06a4308_0
pyparsing=3.0.9=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=2.0.1=py3.10_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=py310h06a4308_0
safetensors=0.3.3=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=py310h1128e8f_0
scipy=1.11.1=py310h5f9d8c6_0
sentry-sdk=1.31.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=68.0.0=py310h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.9.26=pypi_0
timm=0.9.7=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.0.2=py310_cu118
torchmetrics=1.2.0=pypi_0
torchtriton=2.0.0=py310
torchvision=0.15.2=py310_cu118
tqdm=4.65.0=py310h2f386ee_0
typing_extensions=4.7.1=py310h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py310h06a4308_0
wandb=0.15.11=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
